+++ FAManagementStudio/Models/Db/DatabaseInfo.cs
@@ -1,4 +1,5 @@
 ﻿using FAManagementStudio.Common;
+using FAManagementStudio.Models.Common;
 using FirebirdSql.Data.FirebirdClient;
 using System;
 using System.Collections.Generic;
