+++ FAManagementStudio/Models/Db/DomainInfo.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using FAManagementStudio.Models.Common;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
