+++ FAManagementStudio/Models/Db/TableInfo.cs
@@ -1,4 +1,5 @@
 ﻿using FAManagementStudio.Common;
+using FAManagementStudio.Models.Common;
 using FAManagementStudio.Models.Firebird;
 using FirebirdSql.Data.FirebirdClient;
 using System;
